requirements.txt
numpy
protobuf<3.20,>=3.9.2
h5py
protobuf~=3.19.0

keras==2.9.0

opencv-python-headless==4.6.0.66
Pillow==9.2.0
streamlit==1.11.1
tensorflow==2.9.1

matplotlib==3.3.2
matplotlib.pyplot
